+++ l10n/ro.js
@@ -178,6 +178,8 @@ OC.L10N.register(
             "%n vizită",
             "%n vizite",
             "%n de vizite"
-        ]
+        ],
+        "Album link not found": "Linkul de album nu a fost găsit",
+        "This album link does not exist": "Acest link de album nu există"
     },
 "nplurals=3; plural=(n==1 ? 0 : (n==0 || (n%100 > 0 && n%100 < 20)) ? 1 : 2);");